Slimming Bladeburner.ts
* Move getActionIdFromTypeAndName to ActionIdentifier.ts
* Remove parseCommandArguments, reuse parseCommand function from terminal parsing.
